+++ src/Controller/Admin/AppController.php
@@ -45,6 +45,27 @@ abstract class AppController extends BaseAppController
     }
 
     /**
+     * Called after the controller action is run, but before the view is rendered
+     * @param \Cake\Event\Event $event An Event instance
+     * @return void
+     * @since 2.27.5
+     */
+    public function beforeRender(Event $event)
+    {
+        parent::beforeRender($event);
+
+        //For `index` action, saves request target and current controller name
+        //  in session as a referer. The controller name will be used to verify
+        $request = $this->getRequest();
+        if ($request->isAction('index')) {
+            $request->getSession()->write('referer', [
+                'controller' => $request->getParam('controller'),
+                'target' => $request->getRequestTarget(),
+            ]);
+        }
+    }
+
+    /**
      * Initialization hook method
      * @return void
      */
@@ -53,6 +74,31 @@ abstract class AppController extends BaseAppController
         parent::initialize();
 
         $this->loadComponent('MeTools.Uploader');
+    }
+
+    /**
+     * Returns the referring URL for this request.
+     *
+     * Unlike the original method, this can return the `index` action of the same
+     *  controller (if it has been indicated as the `$default` parameter),
+     *  preserving also the query string
+     * @param string|array|null $default Default URL to use if `HTTP_REFERER`
+     *  cannot be read from headers
+     * @param bool $local If `true`, restrict referring URLs to local server
+     * @return string Referring URL
+     * @see beforeRender()
+     * @since 2.27.5
+     */
+    public function referer($default = '/', bool $local = false): string
+    {
+        $request = $this->getRequest();
+        $session = $request->getSession();
+        if ($default == ['action' => 'index'] &&
+            $session->read('referer.controller') == $request->getParam('controller')) {
+            return $session->read('referer.target');
+        }
+
+        return parent::referer($default, $local);
     }
 
     /**