+++ src/Controller/Admin/BannersPositionsController.php
@@ -63,7 +63,7 @@ class BannersPositionsController extends AppController
             if ($this->BannersPositions->save($position)) {
                 $this->Flash->success(I18N_OPERATION_OK);
 
-                return $this->redirect(['action' => 'index']);
+                return $this->redirect($this->referer(['action' => 'index']));
             }
 
             $this->Flash->error(I18N_OPERATION_NOT_OK);
@@ -87,7 +87,7 @@ class BannersPositionsController extends AppController
             if ($this->BannersPositions->save($position)) {
                 $this->Flash->success(I18N_OPERATION_OK);
 
-                return $this->redirect(['action' => 'index']);
+                return $this->redirect($this->referer(['action' => 'index']));
             }
 
             $this->Flash->error(I18N_OPERATION_NOT_OK);
@@ -115,6 +115,6 @@ class BannersPositionsController extends AppController
         }
         call_user_func([$this->Flash, $method], $message);
 
-        return $this->redirect(['action' => 'index']);
+        return $this->redirect($this->referer(['action' => 'index']));
     }
 }