+++ src/Controller/Admin/UsersGroupsController.php
@@ -63,7 +63,7 @@ class UsersGroupsController extends AppController
             if ($this->UsersGroups->save($group)) {
                 $this->Flash->success(I18N_OPERATION_OK);
 
-                return $this->redirect(['action' => 'index']);
+                return $this->redirect($this->referer(['action' => 'index']));
             }
 
             $this->Flash->error(I18N_OPERATION_NOT_OK);
@@ -87,7 +87,7 @@ class UsersGroupsController extends AppController
             if ($this->UsersGroups->save($group)) {
                 $this->Flash->success(I18N_OPERATION_OK);
 
-                return $this->redirect(['action' => 'index']);
+                return $this->redirect($this->referer(['action' => 'index']));
             }
 
             $this->Flash->error(I18N_OPERATION_NOT_OK);
@@ -117,6 +117,6 @@ class UsersGroupsController extends AppController
         }
         call_user_func([$this->Flash, $method], $message);
 
-        return $this->redirect(['action' => 'index']);
+        return $this->redirect($this->referer(['action' => 'index']));
     }
 }