+++ templates/Admin/Photos/index_as_grid.php
@@ -21,43 +21,53 @@ $this->extend('/Admin/common/BannersAndPhotos/index');
             <div class="card">
                 <ul class="list-group list-group-flush">
                     <li class="list-group-item p-1 text-truncate text-center">
-                        <?= $this->Html->link($photo->filename, ['action' => 'edit', $photo->id]) ?>
+                        <?= $this->Html->link($photo->get('filename'), ['action' => 'edit', $photo->get('id')]) ?>
                     </li>
                     <li class="list-group-item p-1 small text-center">
-                        <samp><?= I18N_ID ?> <?= $photo->id ?></samp>
+                        <samp><?= I18N_ID ?> <?= $photo->get('id') ?></samp>
                     </li>
                     <li class="list-group-item p-1 small text-center">
                         <?= __d('me_cms', 'Album') ?>:
                         <?= $this->Html->link(
-                            $photo->album->title,
-                            ['?' => ['album' => $photo->album->id]],
+                            $photo->get('album')->get('title'),
+                            ['?' => ['album' => $photo->get('album')->get('id')]],
                             ['title' => I18N_BELONG_ELEMENT]
                         ) ?>
                     </li>
                     <li class="list-group-item p-1 small text-center">
-                        (<?= $photo->created->i18nFormat() ?>)
+                        (<?= $photo->get('created')->i18nFormat() ?>)
                     </li>
                 </ul>
 
                 <?php
-                echo $this->Thumb->fit($photo->path, ['width' => 400], ['class' => 'card-img-bottom']);
+                echo $this->Thumb->fit($photo->get('path'), ['width' => 400], ['class' => 'card-img-bottom']);
 
-                $actions = [
-                    $this->Html->button(null, ['action' => 'edit', $photo->id], [
-                        'class' => 'btn-link',
-                        'icon' => 'pencil-alt',
-                        'title' => I18N_EDIT,
-                    ]),
-                    $this->Html->button(null, ['action' => 'download', $photo->id], [
-                        'class' => 'btn-link',
-                        'icon' => 'download',
-                        'title' => I18N_DOWNLOAD,
-                    ]),
-                ];
+                $actions = [];
+
+                //If Fancybox is enabled, adds the preview action
+                if (getConfig('default.fancybox')) {
+                    $actions[] = $this->Html->button(null, ['action' => 'edit', $photo->get('id')], [
+                        'class' => 'btn-link fancybox',
+                        'icon' => 'search',
+                        'title' => I18N_PREVIEW,
+                        'data-fancybox-href' => $this->Thumb->resizeUrl($photo->get('path'), ['height' => 1280]),
+                    ]);
+                }
+
+                $actions[] = $this->Html->button(null, ['action' => 'edit', $photo->get('id')], [
+                    'class' => 'btn-link',
+                    'icon' => 'pencil-alt',
+                    'title' => I18N_EDIT,
+                ]);
+                $actions[] = $this->Html->button(null, ['action' => 'download', $photo->get('id')], [
+                    'class' => 'btn-link',
+                    'icon' => 'download',
+                    'title' => I18N_DOWNLOAD,
+                ]);
 
                 //Only admins and managers can delete photos
                 if ($this->Auth->isGroup(['admin', 'manager'])) {
-                    $actions[] = $this->Form->postButton(null, ['action' => 'delete', (string)$photo->id], [
+                    $actions[] = $this->Form->postButton(null, ['action' => 'delete', (string)$photo->get('id')], [
                         'class' => 'btn-link text-danger',
                         'icon' => 'trash-alt',
                         'title' => I18N_DELETE,
@@ -69,8 +79,8 @@ $this->extend('/Admin/common/BannersAndPhotos/index');
                 if ($photo->active) {
                     $actions[] = $this->Html->button(null, [
                         '_name' => 'photo',
-                        'slug' => $photo->album->slug,
-                        'id' => (string)$photo->id,
+                        'slug' => $photo->get('album')->get('slug'),
+                        'id' => (string)$photo->get('id'),
                     ], [
                         'class' => 'btn-link',
                         'icon' => 'external-link-alt',
@@ -78,7 +88,7 @@ $this->extend('/Admin/common/BannersAndPhotos/index');
                         'title' => I18N_OPEN,
                     ]);
                 } else {
-                    $actions[] = $this->Html->button(null, ['_name' => 'photosPreview', $photo->id], [
+                    $actions[] = $this->Html->button(null, ['_name' => 'photosPreview', $photo->get('id')], [
                         'class' => 'btn-link',
                         'icon' => 'external-link-alt',
                         'target' => '_blank',