+++ templates/element/admin/sidebar.php
@@ -22,7 +22,7 @@ use MeTools\Core\Plugin;
     echo $this->MenuBuilder->renderAsCollapse('MeCms', 'sidebar-accordion');
 
     foreach (Plugin::all(['exclude' => ['MeCms', 'MeTools', 'Assets', 'DatabaseBackup', 'Thumber']]) as $plugin) {
-        $menus = $this->MenuBuilder->renderAsCollapse($plugin);
+        $menus = $this->MenuBuilder->renderAsCollapse($plugin, 'sidebar-accordion');
         if ($menus) {
             echo $this->Html->h6($plugin);
             echo $menus;